move various misc components into a components/ folder
