+++ src/components/NavItems/Assistant/AssistantScrapeResults/AssistantTextSpanClassification.jsx
@@ -19,8 +19,8 @@ import { hexToRgb } from "@mui/material/styles";
 
 import HelpOutlineOutlinedIcon from "@mui/icons-material/HelpOutlineOutlined";
 
+import { ThresholdSlider } from "@/components/NavItems/Assistant/components";
 import {
-  ThresholdSlider,
   getPersuasionCategoryTechnique,
   mergeSpanIndices,
   rgbToLuminance,
