+++ src/components/NavItems/Assistant/AssistantScrapeResults/AssistantWarnings.jsx
@@ -25,7 +25,7 @@ import {
   TransKinitAuthor,
   TransOntotextAuthor,
   TransPrevFactChecksLink,
-} from "../TransComponents";
+} from "../components";
 
 const AssistantWarnings = () => {
   const keyword = i18nLoadNamespace("components/NavItems/tools/Assistant");
