+++ src/components/NavItems/Assistant/components/TransComponents.jsx
@@ -3,7 +3,7 @@ import { Trans } from "react-i18next";
 
 import Chip from "@mui/material/Chip";
 
-import { STANCE_CATEGORIES, STANCE_COLOR_MAP } from "./constants";
+import { STANCE_CATEGORIES, STANCE_COLOR_MAP } from "../constants";
 
 const URLS = {
   SUPPORTED_TOOLS:
