+++ src/components/NavItems/Assistant/components/index.js
@@ -1,0 +1,3 @@
+export { ThresholdSlider } from "./ThresholdSlider";
+export { DomainDialog } from "./DomainDialog";
+export * from "./TransComponents";
