+++ src/components/NavItems/Assistant/utils/index.js
@@ -26,6 +26,5 @@ export {
   mergeSpanIndices,
   getPersuasionCategoryTechnique,
   createGaugeChart,
-  ThresholdSlider,
   scrollToElement,
 } from "./textUtils";
