+++ src/components/NavItems/Assistant/utils/textUtils.jsx
@@ -2,9 +2,6 @@ import React from "react";
 import GaugeChart from "react-gauge-chart";
 
 import Box from "@mui/material/Box";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import Slider from "@mui/material/Slider";
 import Typography from "@mui/material/Typography";
 
 import _ from "lodash";
@@ -274,59 +271,6 @@ export function createGaugeChart(
         </Typography>
       </Box>
     </>
-  );
-}
-
-/**
- * Slider component utilised by news framing, news genre, persuasion techniques and subjectivity
- * @param { importantSentenceThreshold, handleSliderChange, keyword }
- * @returns slider component
- */
-export function ThresholdSlider({
-  importantSentenceThreshold,
-  handleSliderChange,
-  keyword,
-}) {
-  const marks = [
-    {
-      value: 0,
-      label: keyword("threshold_slider_low"),
-    },
-    {
-      value: 99,
-      label: keyword("threshold_slider_high"),
-    },
-  ];
-
-  /**
-   * Scales slider threshold to change from range 0 to 100 to 0 to 1
-   * @param value
-   * @returns number between 0 and 1
-   */
-  const scaleValue = (value) => {
-    return value / 100;
-  };
-
-  return (
-    <List>
-      <ListItem>
-        <Slider
-          aria-label="important sentence threshold slider"
-          marks={marks}
-          step={1}
-          min={0}
-          max={99}
-          scale={scaleValue}
-          value={importantSentenceThreshold}
-          onChange={handleSliderChange}
-          sx={{
-            "& .MuiSlider-markLabel": {
-              fontSize: "small",
-            },
-          }}
-        />
-      </ListItem>
-    </List>
   );
 }
